problema 10

+++ laboratorio_2/problemas.h
@@ -16,7 +16,7 @@ char * problema6(char *c);
 void problema7(char *c);
 void problema8(char *c1, char *c2);
 long long int problema9(char *c, long long int n);
-long long int problema10(char *c);
+int problema10(char *c);
 void problema11(char c, char *c2);
 bool problema12(int *mat, int elementos);
 int problema13(int *mat);
